REFACTOR reused assertion functions to `PdfconduitTestCase`
